Revert "Issue#1 scaling"
