+++ main.py
@@ -6,17 +6,20 @@ import random
 import pygame
 from pygame.locals import *
 
-import helper_func as hf
-
 pygame.init()
 
 FORMATS = ['jpg', 'jpeg', 'png', 'gif', 'bmp', 'pcx', 'tga', 'tif', 'lbm',
            'pbm', 'pgm', 'ppm', 'xpm']
 
-# Obtain the current display info for scaling of the image
-# see line below pygame.transform.scale
-infoObject = pygame.display.Info()
-SCREEN_RES = (infoObject.current_w, infoObject.current_h)
+def list_img_paths(cpath):
+    """
+    Lists all image filenames in the path directory, p.
+    curpath: String: Correctly formatted path according to platform
+    Returns:
+    List of strings of image filenames.
+    """
+
+    return [i for i in listdir(cpath) if i.split('.')[-1].lower() in FORMATS]
 
 
 class SlideShow:
@@ -25,21 +28,18 @@ class SlideShow:
     Runs slideshow using pygame
     """
 
-    def __init__(self, time_delay=3000, path2imgs=getcwd()):
-        # Sorting out the images
-        self.imgsrc = hf.list_img_paths(path2imgs)
-        self.imgpath = random.choice(self.imgsrc)
-        # will be populated as pygame.Surface in run method
-        self.img = None
-
-        # Defines how much time passes between image switch
-        self.time_delay = time_delay
+    def __init__(self, imgpath=getcwd()):
+        self.imgsrc = list_img_paths(imgpath)
+        self.img = random.choice(self.imgsrc)
+        self.imgpath = imgpath
 
 
     def run(self):
         """
         DEFINITION
         TODOs
+        * Set up the github for version control and project structure
+        * make image scaled to screen
         * Deal with rotation images (orientation landscape vs. portrait)
             * Most likely decide on landscape or portrait and pick a good screen.
         * Switch the image and regular intervals (X Minutes)
@@ -47,28 +47,27 @@ class SlideShow:
         * The sleep function for Raspberry and wake up
         """
 
-        black = (0, 0, 0)
+        # Make image now scaled to surface and resizable acording to screen size
+        # https://stackoverflow.com/questions/20002242/how-to-scale-images-to-screen-size-in-pygame
 
-        # Chooses the size of your display
-        # Makes it portable to varying screens
+        img = pygame.image.load(self.img)
+
+        white = (0, 0, 0)
         size = (0, 0)
         screen = pygame.display.set_mode(size, FULLSCREEN)
-
+        screen.fill((white))
         running = True
 
         while running:
-            self.imgpath = random.choice(self.imgsrc)
-            self.img = pygame.image.load(self.imgpath)
+            new_img = random.choice(self.imgsrc)
 
-            # Scale the image according to screen resolution
-            new_dim = hf.aspect_scale(self.img.get_size(), SCREEN_RES)
-            self.img = pygame.transform.scale(self.img, new_dim)
-            # get the coordinates to center image
-            centre_coord = hf.center_img(self.img.get_size(), SCREEN_RES)
+            screen.fill((white))
 
-            screen.fill(black)
-            screen.blit(self.img, centre_coord)
-            pygame.time.delay(self.time_delay)
+            screen.blit(img, (0, 0))
+            self.img = pygame.image.load(new_img)
+            screen.fill(white)
+            screen.blit(self.img, (0, 0))
+            pygame.time.delay(3000)
             pygame.display.flip()
 
             # event handling, gets all event from the event queue
